[wip] refactored non-command methods within redisai.c into dag, run_info, background_workers, and model_script_run_session
