+++ src/redisai.h
@@ -4,6 +4,9 @@
 #include <stdbool.h>
 #include "redismodule.h"
 #include "util/dict.h"
+#include "model_struct.h"
+#include "model_script_run_session.h"
+#include "background_workers.h"
 
 #define REDISAI_LLAPI_VERSION 1
 
@@ -26,7 +29,6 @@ typedef struct RAI_Error RAI_Error;
 
 #define REDISAI_DEVICE_CPU 0
 #define REDISAI_DEVICE_GPU 1
-#define REDISAI_DEFAULT_THREADS_PER_QUEUE 1
 
 #define REDISAI_ERRORMSG_PROCESSING_ARG "ERR: error processing argument"
 #define REDISAI_ERRORMSG_THREADS_PER_QUEUE "ERR: error setting THREADS_PER_QUEUE to"
@@ -36,20 +38,6 @@ enum RedisAI_DataFmt {
   REDISAI_DATA_BLOB = 0,
   REDISAI_DATA_VALUES,
   REDISAI_DATA_NONE
-};
-
-struct RedisAI_RunInfo {
-  RedisModuleBlockedClient *client;
-  RedisModuleString *runkey;
-  RedisModuleString **outkeys;
-  RAI_ModelRunCtx *mctx;
-  RAI_ScriptRunCtx *sctx;
-  int status;
-  long long duration_us;
-  RAI_Error* err;
-  int use_local_context;
-  AI_dict *dagTensorsContext;
-  AI_dict *dagTensorsPersistentContext;
 };
 
 RAI_Tensor* MODULE_API_FUNC(RedisAI_TensorCreate)(const char* dataTypeStr, long long* dims, int ndims);
